[#3] refactor: make response package
