+++ src/main/java/sparta/com/sappun/domain/sample/controller/SampleController.java
@@ -17,7 +17,7 @@ import sparta.com.sappun.domain.sample.dto.response.SampleGetRes;
 import sparta.com.sappun.domain.sample.dto.response.SampleSaveRes;
 import sparta.com.sappun.domain.sample.dto.response.SampleUpdateRes;
 import sparta.com.sappun.domain.sample.service.SampleService;
-import sparta.com.sappun.global.CommonResponse;
+import sparta.com.sappun.global.response.CommonResponse;
 
 @RestController
 @RequestMapping("/api/samples")
